requirements.txt
sklearn
scipy
nltk
langdetect
numpy
pandas
matplotlib
Flask
flask_restplus
flask_wtf
fuzzywuzzy
invoke
requests
python-dotenv
wtforms
requests
gensim
